+++ src/main/java/com/yuuko/core/events/controllers/GenericGuildController.java
@@ -2,8 +2,8 @@ package com.yuuko.core.events.controllers;
 
 import com.yuuko.core.Cache;
 import com.yuuko.core.Configuration;
-import com.yuuko.core.Metrics;
 import com.yuuko.core.database.DatabaseFunctions;
+import com.yuuko.core.metrics.Metrics;
 import com.yuuko.core.modules.core.commands.CommandSetup;
 import com.yuuko.core.utilities.MessageHandler;
 import com.yuuko.core.utilities.Utils;
@@ -40,7 +40,7 @@ public class GenericGuildController {
                         .addField("Servers", Metrics.GUILD_COUNT + "", true)
                         .addField("Commands", Cache.COMMANDS.size() + "", true)
                         .addField("Invocation", Configuration.GLOBAL_PREFIX + ", `" + Utils.getServerPrefix(e.getGuild().getId()) + "`", true)
-                        .addField("Uptime", Metrics.RUNTIME.toString(), true)
+                        .addField("Uptime", Metrics.UPTIME + "", true)
                         .addField("Ping", Metrics.PING + "", true);
                 MessageHandler.sendMessage(textChannel, about.build());
             });