+++ src/main/java/com/yuuko/core/metrics/Metrics.java
@@ -1,14 +1,14 @@
-package com.yuuko.core;
+package com.yuuko.core.metrics;
 
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 
 public class Metrics {
-    public static final AtomicInteger RUNTIME = new AtomicInteger();
+    public static long UPTIME = 0;
+    public static long MEMORY_TOTAL = 0;
+    public static long MEMORY_USED = 0;
+    public static int GUILD_COUNT = 0;
     public static final AtomicLong PING = new AtomicLong();
-    public static int DB_POOL_ACTIVE;
-    public static int DB_POOL_IDLE;
-    public static int GUILD_COUNT;
     public static final AtomicInteger MESSAGES_PROCESSED = new AtomicInteger();
     public static final AtomicInteger REACTS_PROCESSED = new AtomicInteger();
     public static final AtomicInteger COMMANDS_PROCESSED = new AtomicInteger();