+++ src/main/java/com/yuuko/core/scheduler/ScheduleHandler.java
@@ -7,7 +7,6 @@ import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
 
 public class ScheduleHandler {
-
     private static final Set<ScheduledFuture<?>> tasks = new HashSet<>();
     private static final ScheduledExecutorService schedulerService = Executors.newScheduledThreadPool(1);