+++ src/main/java/com/yuuko/core/scheduler/tasks/UpdateMetricsTask.java
@@ -1,13 +1,16 @@
 package com.yuuko.core.scheduler.tasks;
 
-import com.yuuko.core.Cache;
-import com.yuuko.core.Metrics;
+import com.yuuko.core.metrics.Metrics;
 import com.yuuko.core.scheduler.Task;
+
+import java.lang.management.ManagementFactory;
 
 public class UpdateMetricsTask implements Task {
 
     @Override
     public void handle() {
-        Metrics.PING.set(Cache.JDA.getPing());
+        Metrics.UPTIME = ManagementFactory.getRuntimeMXBean().getUptime();
+        Metrics.MEMORY_TOTAL = Runtime.getRuntime().totalMemory();
+        Metrics.MEMORY_USED = (Runtime.getRuntime().totalMemory() - Runtime.getRuntime().freeMemory());
     }
 }