+++ src/main/java/com/yuuko/core/scheduler/tasks/UploadShardMetricsTask.java
@@ -3,11 +3,11 @@ package com.yuuko.core.scheduler.tasks;
 import com.yuuko.core.database.DatabaseFunctions;
 import com.yuuko.core.scheduler.Task;
 
-public class UpdateServerStatusTask implements Task {
+public class UploadShardMetricsTask implements Task {
 
     @Override
     public void handle() {
-        new DatabaseFunctions().updateServerStatus();
+        new DatabaseFunctions().updateShardMetrics();
     }
 
 }