+++ src/main/java/com/yuuko/core/utilities/Utils.java
@@ -2,8 +2,8 @@ package com.yuuko.core.utilities;
 
 import com.yuuko.core.Cache;
 import com.yuuko.core.Configuration;
-import com.yuuko.core.Metrics;
 import com.yuuko.core.database.DatabaseFunctions;
+import com.yuuko.core.metrics.Metrics;
 import net.dv8tion.jda.core.EmbedBuilder;
 import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.entities.*;
@@ -175,7 +175,7 @@ public final class Utils {
         System.out.println("┃ " + Cache.LAST_THIRTEEN.get(11));
         System.out.println("┃ " + Cache.LAST_THIRTEEN.get(12));
         System.out.println("┣━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━[STATISTICS]━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┫");
-        System.out.println("┃ Uptime: " + Metrics.RUNTIME + ", Ping: " + Metrics.PING + ", DB Idle: " + Metrics.DB_POOL_IDLE + ", DB Active: " + Metrics.DB_POOL_ACTIVE);
+        System.out.println("┃ Uptime: " + Metrics.UPTIME + ", Ping: " + Metrics.PING);
         System.out.println("┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛");
     }